LayoutTests/media/media-source/worker/media-managedmse-tracks-worker.html

<!DOCTYPE html> <!-- webkit-test-runner [ ManagedMediaSourceEnabled=true MediaSourceEnabled=true ] -->
<html>
<head>
    <title>managedmediasource in worker</title>
    <script src="../../../media/video-test.js"></script>
    <script>
    window.addEventListener('load', async event => {
        findMediaElement();

        testExpected('ManagedMediaSource.canConstructInDedicatedWorker', true);
        const canPlayPromise = waitFor(video, 'canplay', true);
        const worker = new Worker('worker.js');
        worker.onmessage = async (msg) => {
            switch (msg.data.topic) {
            case 'handle':
                logResult(true, 'received handle message: ' + msg.data.arg);
                video.srcObject = msg.data.arg;
                break;
            case 'info':
                logResult(true, 'info message from worker: ' + msg.data.arg);
                break;
            case 'done':
                await canPlayPromise;
                testExpected('video.readyState', HTMLMediaElement.HAVE_CURRENT_DATA, '>=');
                testExpected('video.buffered.length', 1);
                testExpected('video.audioTracks.length', 1);
                testExpected('video.audioTracks.item(0).language', "eng");
                testExpected('video.videoTracks.length', 1);
                testExpected('video.videoTracks.item(0).language', "eng");
                endTest();
                break;
            case 'error':
                logResult(false, 'error message from worker: ' + msg.data.arg);
                endTest();
                break;
            default:
                logResult(false, 'error: Unrecognized topic in message from worker');
                break;
            }
        };
        run('video.disableRemotePlayback = true');
        worker.postMessage({topic: 'setup'});
        worker.postMessage({topic: 'manifest', arg: '../content/test-fragmented-manifest.json'});
    });
    </script>
</head>
<body>
    <video controls></video>
</body>
</html>
